refactor: reorganize agent implementation filenames
- Renamed agent implementation files to have consistent class_ prefix
- Updated imports in __init__.py and execute.py to match new filenames
- Fixed import in agent_class_merge.py
