+++ coauthor/agent/__init__.py
@@ -13,9 +13,9 @@ from .tool_state import ToolState
 from .agent_base import BaseReflectionAgent
 
 # Specialized agent implementations
-from .agent_cot import CoTAgent
-from .agent_direct import DirectAgent
-from .agent_merge import AgentMerge
+from .agent_class_cot import CoTAgent
+from .agent_class_direct import DirectAgent
+from .agent_class_merge import AgentMerge
 
 # Utility and loading functions
 from .agent_load import load_agent_settings_and_prompts
