+++ coauthor/agent/agent_class_merge.py
@@ -5,7 +5,7 @@ from ..logger import logger
 
 from .agent_config import AgentConfig
 from .agent_dataclass import AgentSettings, AgentPrompts
-from .agent_direct import DirectAgent
+from .agent_class_direct import DirectAgent
 from .agent_state import AgentStateRound, AgentStateGlobal
 from .model_handler import ModelHandler
