+++ coauthor/execute.py
@@ -12,14 +12,13 @@ from .agent.agent_dataclass import AgentSettings, AgentPrompts
 from .agent.agent_config import AgentConfig
 from .agent.agent_load import load_agent_settings_and_prompts
 
-from .agent.agent_base import BaseReflectionAgent
-from .agent.agent_cot import CoTAgent
-from .agent.agent_direct import DirectAgent
-from .agent.agent_merge import AgentMerge
-
 from .agent.model_registry import MODEL_CONFIGS
 from .agent.model_factory import ModelFactory
 
+from .agent.agent_base import BaseReflectionAgent
+from .agent.agent_class_cot import CoTAgent
+from .agent.agent_class_direct import DirectAgent
+from .agent.agent_class_merge import AgentMerge
 
 load_dotenv()
